refactor: decouple search policy services
