+++ src/api_query/src/search/multi_streams.rs
@@ -804,8 +804,26 @@ pub async fn _search_partition_multi(
         }
     }
 
-    let search_fut =
-        SearchService::search_partition_multi(&trace_id, &org_id, user_id, stream_type, &req);
+    let mut max_query_ranges = Vec::with_capacity(req.sql.len());
+    for sql in &req.sql {
+        let stream_names = resolve_stream_names(sql).unwrap_or_default();
+        max_query_ranges.push(
+            crate::common::utils::stream::get_max_query_range(
+                &stream_names,
+                &org_id,
+                user_id,
+                stream_type,
+            )
+            .await,
+        );
+    }
+    let search_fut = SearchService::search_partition_multi(
+        &trace_id,
+        &org_id,
+        &max_query_ranges,
+        stream_type,
+        &req,
+    );
     let search_res = if cfg.common.should_create_span() {
         search_fut.instrument(http_span).await
     } else {
@@ -1486,6 +1504,20 @@ pub async fn search_multi_stream(
     #[cfg(not(feature = "enterprise"))]
     let audit_ctx = None;
 
+    let mut max_query_ranges = Vec::with_capacity(queries.len());
+    for req in &queries {
+        let stream_names = resolve_stream_names(&req.query.sql).unwrap_or_default();
+        max_query_ranges.push(
+            crate::common::utils::stream::get_max_query_range(
+                &stream_names,
+                &org_id,
+                &user_id,
+                stream_type,
+            )
+            .await,
+        );
+    }
+
     // Spawn the multi-stream search task.
     // Pass needs_post_vrl (true only when per_query_response=true AND ResultArray VRL) and
     // the top-level query_fn for post-hoc application. When needs_post_vrl is false,
@@ -1495,6 +1527,7 @@ pub async fn search_multi_stream(
         user_id,
         trace_id.clone(),
         queries,
+        max_query_ranges,
         stream_type,
         http_span,
         tx,
