+++ src/api_query/src/search/search_stream.rs
@@ -553,6 +553,13 @@ pub async fn search_http2_stream(
     #[cfg(not(feature = "enterprise"))]
     let audit_ctx = None;
     let search_type = req.search_type;
+    let max_query_range = crate::common::utils::stream::get_max_query_range(
+        &stream_names,
+        &org_id,
+        &user_id,
+        stream_type,
+    )
+    .await;
 
     // Spawn the search task in a separate task
     tokio::spawn(process_search_stream_request(
@@ -562,6 +569,7 @@ pub async fn search_http2_stream(
         req,
         stream_type,
         stream_names,
+        max_query_range,
         req_order_by,
         http_span,
         tx,
@@ -944,6 +952,13 @@ pub async fn values_http2_stream(
 
     // Pattern extraction is not supported for values endpoint
     let extract_patterns = false;
+    let max_query_range = crate::common::utils::stream::get_max_query_range(
+        &stream_names,
+        &org_id,
+        &user_id,
+        stream_type,
+    )
+    .await;
 
     // Spawn the search task to process the request
     tokio::spawn(process_search_stream_request(
@@ -953,6 +968,7 @@ pub async fn values_http2_stream(
         req,
         stream_type,
         stream_names,
+        max_query_range,
         OrderBy::default(),
         http_span,
         tx,
