+++ src/audit/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "audit"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+publish = false
+
+[features]
+default = []
+enterprise = ["dep:o2_enterprise"]
+
+[dependencies]
+anyhow.workspace = true
+config.workspace = true
+log.workspace = true
+o2_enterprise = { workspace = true, optional = true, default-features = false }
+proto.workspace = true
+tokio.workspace = true
