+++ src/core/src/lib.rs
@@ -69,7 +69,6 @@ pub mod runtime_metrics;
 pub mod schema;
 mod schema_watcher;
 pub use search_service as search;
-mod search_hooks;
 #[cfg(feature = "enterprise")]
 pub mod search_jobs;
 pub mod self_reporting;
