+++ src/search_service/Cargo.toml
@@ -8,16 +8,19 @@ publish = false
 [features]
 default = []
 enterprise = [
+    "audit/enterprise",
     "dep:o2_enterprise",
     "common/enterprise",
     "config/enterprise",
     "promql/enterprise",
     "search/enterprise",
+    "usage_reporting/enterprise",
 ]
 vectorscan = ["dep:o2_enterprise", "o2_enterprise/vectorscan", "search/vectorscan"]
 vortex = ["dep:vortex", "config/vortex", "o2_enterprise?/vortex", "search/vortex"]
 
 [dependencies]
+audit.workspace = true
 common.workspace = true
 transform.workspace = true
 promql.workspace = true
@@ -65,6 +68,7 @@ tracing-opentelemetry.workspace = true
 url.workspace = true
 vortex = { workspace = true, optional = true }
 wal.workspace = true
+usage_reporting.workspace = true
 
 [dev-dependencies]
 expect-test.workspace = true
