+++ src/search_service/src/grpc_server.rs
@@ -134,7 +134,7 @@ impl Search for Searcher {
         let ret = SearchService::search_partition(
             &req.trace_id,
             &req.org_id,
-            None,
+            req.max_query_range,
             stream_type,
             &request,
             req.skip_max_query_range,
