+++ src/search_service/src/streaming/mod.rs
@@ -21,6 +21,8 @@
 use std::{collections::HashMap, ops::ControlFlow, time::Instant};
 
 use ::search::{AuditContext, SearchResultType};
+#[cfg(feature = "enterprise")]
+use audit::audit;
 use config::{
     cluster::LOCAL_NODE,
     meta::{
@@ -54,11 +56,8 @@ use transform::vector_enrichment::TableRegistry;
 use crate::service::search::sql::visitor::histogram_interval::{
     convert_histogram_interval_to_seconds, generate_histogram_interval,
 };
-#[cfg(feature = "enterprise")]
-use crate::hooks::audit;
 use crate::{
     cache as search_cache,
-    hooks::get_max_query_range,
     inspector::{SearchInspectorFieldsBuilder, search_inspector_fields},
     sql::visitor::histogram_interval::{
         HistogramIntervalVisitor, validate_and_adjust_histogram_interval,
@@ -102,6 +101,7 @@ pub async fn process_search_stream_request(
     mut req: Request,
     stream_type: StreamType,
     stream_names: Vec<String>,
+    max_query_range: i64,
     req_order_by: OrderBy,
     search_span: tracing::Span,
     sender: mpsc::Sender<Result<StreamResponses, infra::errors::Error>>,
@@ -266,8 +266,6 @@ pub async fn process_search_stream_request(
     }
 
     req.query.query_fn = query_fn.clone();
-
-    let max_query_range = get_max_query_range(&stream_names, &org_id, &user_id, stream_type).await; // hours
 
     // HACK: always search from the first partition, this is because to support pagination in http2
     // streaming we need context of no of hits per partition, which currently is not available.
@@ -831,6 +829,7 @@ pub async fn process_search_stream_request_multi(
     user_id: String,
     trace_id: String,
     queries: Vec<Request>,
+    max_query_ranges: Vec<i64>,
     stream_type: StreamType,
     search_span: tracing::Span,
     sender: mpsc::Sender<Result<config::meta::search::StreamResponses, infra::errors::Error>>,
@@ -877,6 +876,10 @@ pub async fn process_search_stream_request_multi(
     let mut query_tasks = Vec::new();
 
     for (query_index, mut req) in queries.into_iter().enumerate() {
+        let max_query_range = max_query_ranges
+            .get(query_index)
+            .copied()
+            .unwrap_or_default();
         let query_trace_id = format!("{trace_id}-q{query_index}");
         let org_id_clone = org_id.clone();
         let user_id_clone = user_id.clone();
@@ -930,6 +933,7 @@ pub async fn process_search_stream_request_multi(
                 req,
                 stream_type,
                 stream_names,
+                max_query_range,
                 OrderBy::default(),
                 search_span_clone.clone(),
                 query_sender,
